requirements.txt
requests==2.18.4
discord.py==0.16.12
peewee==3.7.1
discord==0.0.2
PyYAML==3.13
